requirements.txt
# ============================================================
# Dépendances du projet — Knowledge Graph & RAG Assistant
# ============================================================

# --- Web sémantique & RDF ---
rdflib==7.0.0
owlready2==0.46
SPARQLWrapper==2.0.0

# --- Knowledge Graph Embedding ---
pykeen==1.10.2
torch>=2.0.0

# --- NLP ---
spacy==3.7.4
# Télécharger le modèle anglais après installation :
#   python -m spacy download en_core_web_sm

# --- Scraping & APIs ---
requests==2.31.0
beautifulsoup4==4.12.3

# --- Data ---
pandas==2.2.1
numpy==1.26.4

# --- Visualisation & UI ---
matplotlib==3.8.4
scikit-learn==1.4.2
streamlit>=1.30.0

# --- LLM & RAG ---
openai==1.25.0
anthropic==0.26.0
python-dotenv==1.0.1

# --- Notebooks ---
jupyter==1.0.0
ipykernel==6.29.4
